02_django_api/requirements.txt
Django==3.2
django-debug-toolbar==3.4.0
django-split-settings==1.1.0
flake8==4.0.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
